getting into chapter 6
for these chapters I'm using one worksheet because the examples are lessons that cna be so contained. this one is about classes and immutable objects taking the functional approach in scala. I will go back to types at a later point, a lot of that chapter is apparently review for java developers so I should find a way to only take what I need from it.